Hidden-Technical-Debt-Behind-AI-Agents/requirements.txt
beautifulsoup4==4.13.4
chromadb==1.0.9
gradio==5.30.0
ipykernel==6.29.5
kubernetes==32.0.1
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
numpy==2.2.6
openai==1.81.0
pandas==2.2.3
pypdf==5.5.0
pyprojroot==0.3.0
python-dotenv==1.1.0
psycopg[binary]==3.2.9
requests==2.32.3
SQLAlchemy==2.0.41

